fix : @Value 추가, add/remove 메서드 로직 수정

+++ src/main/java/gift/Controller/WishListController.java
@@ -1,24 +1,23 @@
 package gift.Controller;
 
-import gift.Annotation.WishListResolver;
+import gift.Annotation.LoginMemberResolver;
 import gift.Model.User;
 import gift.Model.WishListItem;
 import gift.Service.UserService;
 import gift.Service.WishListService;
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jwts;
-import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.util.WebUtils;
 
 import java.util.List;
 
 @Controller
 public class WishListController {
+    @Value("${jwt.secretKey}")
+    private String secretKey;
     private final WishListService wishlistService;
     private final UserService userService;
 
@@ -29,32 +28,39 @@ public class WishListController {
     }
 
     @GetMapping("/wishlist")
-    public String getWishlist(Model model, HttpServletRequest request) {
+    public String getWishlist(@LoginMemberResolver User user, Model model, HttpServletRequest request) {
+        /*
         Cookie jwtCookie = WebUtils.getCookie(request, "token");
         String token = jwtCookie.getValue();
         Claims claims = Jwts.parser()
-                .setSigningKey(getSigningKey().getBytes())
+                .setSigningKey(secretKey.getBytes())
                 .parseClaimsJws(token)
                 .getBody();
         String userEmail = claims.getSubject();
         User user = userService.findByEmail(userEmail);
+
+         */
         List<WishListItem> wishlist = wishlistService.getWishlist(user.getId());
         model.addAttribute("wishlist", wishlist);
         return "wishlist";
     }
 
-    public String getSigningKey() {
-        return "Yn2kjibddFAWtnPJ2AFlL8WXmohJMCvigQggaEypa5E=";
-    }
-
     @PostMapping("/wishlist/add")
-    public String addWishlistItem(@WishListResolver WishListItem wishlistItem) {
-        wishlistService.addWishlistItem(wishlistItem);
+    public String addWishlistItem(@LoginMemberResolver User user, @RequestBody WishListItem wishListItem) {
+        if(user == null) {
+            return "redirect:/login";
+        }
+        wishListItem.setUserId(user.getId());
+        wishlistService.addWishlistItem(wishListItem);
         return "redirect:/products";
     }
 
     @PostMapping("/wishlist/remove/{productId}")
-    public String removeWishlistItem(@WishListResolver WishListItem wishListItem) {
+    public String removeWishlistItem(@LoginMemberResolver User user, @RequestBody WishListItem wishListItem) {
+        if(user == null) {
+            return "redirect:/login";
+        }
+        wishListItem.setUserId(user.getId());
         wishlistService.removeWishlistItem(wishListItem);
         return "redirect:/wishlist";
     }
